Split command-line parser into different file
